Support generic list separator - beta, missing parts
Enable Schedule to have a task name with comma in its name
The separator is a non valid char pipe "|" which can be changed via a constant. The display is a "; " or can be kept ", ". So user doesn't see the pipe char.
In the Schedule name edit dialog, we must still add a check to disable using the Constant SMBSYNC2_LIST_SEPARATOR in the name

We can even define SMBSYNC2_LIST_SEPARATOR = "," to not change anything, while being able to do in future.

Missing part:
- adding the v6 schedule list

Only real limitation is the external scripts using ",":
A transparent alternative is to change separator to "pipe" but keep external scripts with "," when they are imported in SyncService.java part. However, it will not support scripts with "," in task name, like now by the way !